// src/components/ThreeMashAboutUsPage/types.ts
// This file is auto-generated — do not edit manually.
export interface Props {
  introTitleHtml?: string;
  introContentHtml?: string;
  block1Html?: string;
  block1ImageUrl?: string;
  block1ImageAlt?: string;
  block2Html?: string;
  block2ImageUrl?: string;
  block2ImageAlt?: string;
  valuesTitle?: string;
  value1Title?: string;
  value1SubTitle?: string;
  value1ImageUrl?: string;
  value1ImageAlt?: string;
  value2Title?: string;
  value2SubTitle?: string;
  value2ImageUrl?: string;
  value2ImageAlt?: string;
  value3Title?: string;
  value3SubTitle?: string;
  value3ImageUrl?: string;
  value3ImageAlt?: string;
  value4Title?: string;
  value4SubTitle?: string;
  value4ImageUrl?: string;
  value4ImageAlt?: string;
  backgroundColor?: string;
  textColor?: string;
  mutedTextColor?: string;
  panelColor?: string;
  maxWidth?: number;
}